Drop build_runner dep from build_test
Most consumers of build_test will need to test Builder instances. The
testPhases utility is most useful from within the build_runner package.

- Drop dep on build_runner from build_test
- Add dev_dependency on build_barback from build. Previously this was
  picked up transitively build_test -> build_runner -> build_barback
- Move utilities specific to classes in build_runner from build_test
  into build_runner/test/common. This includes duplicating the
  InMemoryReader/Writer so that the version in build_runner can
  implement the extended interfaces
- Clean up unnecessary method implementations for asset reader/writer
  implementations
- Remove duplicated test for BarbackResolvers
- Remove tests from build_test package. Previous test was only useful
  for testPhases
- Clean up signature of testBuilder. It had been copied from testPhases
  but some arguments were unused
