+++ build_test/lib/src/in_memory_reader.dart
@@ -3,14 +3,12 @@
 // BSD-style license that can be found in the LICENSE file.
 import 'dart:async';
 import 'dart:convert';
-import 'dart:mirrors';
 
 import 'package:build/build.dart';
-import 'package:build_runner/build_runner.dart';
 
 import 'in_memory_writer.dart';
 
-class InMemoryAssetReader implements RunnerAssetReader {
+class InMemoryAssetReader implements AssetReader {
   final Map<AssetId, DatedString> assets;
 
   InMemoryAssetReader(this.assets);
@@ -24,38 +22,5 @@ class InMemoryAssetReader implements RunnerAssetReader {
   Future<String> readAsString(AssetId id, {Encoding encoding: UTF8}) async {
     if (!await hasInput(id)) throw new AssetNotFoundException(id);
     return assets[id].value;
-  }
-
-  @override
-  Stream<AssetId> listAssetIds(Iterable<InputSet> inputSets) async* {
-    for (var id in assets.keys) {
-      var matches = inputSets.any((inputSet) {
-        if (inputSet.package != id.package) return false;
-        return inputSet.globs.any((glob) => glob.matches(id.path));
-      });
-      if (matches) yield id;
-    }
-  }
-
-  final Map<Uri, LibraryMirror> _allLibraries = currentMirrorSystem().libraries;
-
-  @override
-  Future<DateTime> lastModified(AssetId id) async {
-    if (!assets.containsKey(id)) {
-      /// Support reading files imported by the test script as well. Also
-      /// pretend like they are very old so that builds don't get invalidated.
-      ///
-      /// Overridden values in [assets] will take precedence.
-      var uri = new Uri(
-          scheme: 'package',
-          path: '${id.package}/${id.path.replaceFirst("lib/", "")}');
-      if (_allLibraries[uri] != null) {
-        return new DateTime.fromMillisecondsSinceEpoch(0);
-      }
-
-      throw new AssetNotFoundException(id);
-    }
-
-    return assets[id].date;
   }
 }
