+++ build_test/lib/src/stub_writer.dart
@@ -10,7 +10,4 @@ class StubAssetWriter implements AssetWriter {
   @override
   Future writeAsString(Asset asset, {Encoding encoding: UTF8}) =>
       new Future.value(null);
-
-  @override
-  Future delete(AssetId id) => new Future.value(null);
 }
