+++ build_test/pubspec.yaml
@@ -1,12 +1,11 @@
 name: build_test
 description: Utilities for writing unit tests of Builders.
-version: 0.2.1
+version: 0.3.0
 author: Dart Team <misc@dartlang.org>
 homepage: https://github.com/dart-lang/build
 
 dependencies:
   build: ^0.6.0
-  build_runner: ^0.1.0
   logging: ^0.11.2
   test: ^0.12.0
   watcher: ^0.9.7
